+++ lib/domain/media/scene.dart
@@ -1,4 +1,5 @@
 import 'package:webfeed_revised/util/iterable.dart';
+import 'package:webfeed_revised/util/xml.dart';
 import 'package:xml/xml.dart';
 
 /// Used to specify a scene of a media element
@@ -13,9 +14,15 @@ class Scene {
   });
 
   /// Parse constructor for the Scene class, used when 'parsing' a feed
-  factory Scene.parse(XmlElement element) => Scene(
-        title: element.findElements('sceneTitle').firstOrNull?.text,
-        description: element.findElements('sceneDescription').firstOrNull?.text,
+  factory Scene.parse(XmlElement element, bool parseHtml) => Scene(
+        title: element
+            .findElements('sceneTitle')
+            .firstOrNull
+            ?.parseText(parseHtml),
+        description: element
+            .findElements('sceneDescription')
+            .firstOrNull
+            ?.parseText(parseHtml),
         startTime: element.findElements('sceneStartTime').firstOrNull?.text,
         endTime: element.findElements('sceneEndTime').firstOrNull?.text,
       );