+++ lib/domain/media/title.dart
@@ -1,3 +1,4 @@
+import 'package:webfeed_revised/util/xml.dart';
 import 'package:xml/xml.dart';
 
 /// The title of a media element
@@ -10,9 +11,9 @@ class Title {
   });
 
   /// Parse constructor for the Title class, used when 'parsing' a feed
-  factory Title.parse(XmlElement element) => Title(
+  factory Title.parse(XmlElement element, bool parseHtml) => Title(
         type: element.getAttribute('type'),
-        value: element.text,
+        value: element.parseText(parseHtml),
       );
 
   /// The type of the title